feat: Add new views for test staff and test management

+++ database/migrations/0001_01_01_000008_create_staff_information.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('user_id');
             $table->string('staff_name',50);
             $table->enum('gender',["M","F","O"]);
-            $table->integer('contact_number');
+            $table->unsignedBigInteger('contact_number');
             $table->string('email');
             $table->text('address');
             $table->string('designation');
